Added wheat and seed placement rules and growth

+++ src/main/java/org/spout/vanilla/material/block/GroundAttachable.java
@@ -49,7 +49,7 @@ public class GroundAttachable extends AbstractAttachable {
 
 	@Override
 	public void setAttachedFace(Block block, BlockFace attachedFace) {
-		block.setData((short) 0);
+		block.setData(0);
 	}
 
 	@Override
